chore(SelectObserver): adjust SelectValueObserver synchronization
